requirements.txt
matplotlib==3.8.0rc1
numpy==1.25.2
openai==0.27.8
Pillow==10.0.0
Requests==2.31.0
st_annotated_text==4.0.0
streamlit==1.25.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.6
wordcloud==1.9.2
~treamlit==1.23.1
